requirements.txt
streamlit
python-dotenv
requests
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-groq
langchain-huggingface
sentence-transformers==2.2.2
torch==2.1.2+cpu --extra-index-url https://download.pytorch.org/whl/cpu
groq